refactor: change directory for some files
