+++ async-race/src/App.ts
@@ -1,5 +1,5 @@
 import './app.scss';
-import { clickHandler } from './utils/click-handler';
+import { clickHandler } from './utils/click-handler/click-handler';
 import { createPage } from './view/view';
 
 async function start(): Promise<void> {
