+++ async-race/src/utils/click-handler/click-handler.ts
@@ -1,9 +1,9 @@
-import { createCarParams } from '../api/create-one-car';
+import { createCarParams } from '../../api/create-one-car';
 import { createOneCar } from './create-cars';
 import { clickPrevNextBtn } from './open-next-prev-page';
 import { clickRemoveCarBtn } from './remove-car';
 import { clickSelectCarBtn } from './rename-car';
-import { increaseNumberCarsInGarage, onOffNextButton, onOffPrevButton } from './utils';
+import { increaseNumberCarsInGarage, onOffNextButton, onOffPrevButton } from '../utils';
 
 function clickGeneralBtn(): void {
   const buttons: NodeListOf<HTMLButtonElement> = document.querySelectorAll('[data-name="general"]');
